Change examples to explicitly use replaceReducer() for hot reloading
