+++ examples/async/index.js
@@ -1,9 +1,14 @@
 import 'babel-core/polyfill';
-
 import React from 'react';
-import Root from './containers/Root';
+import { Provider } from 'react-redux';
+import App from './containers/App';
+import configureStore from './store/configureStore';
+
+const store = configureStore();
 
 React.render(
-  <Root />,
+  <Provider store={store}>
+    {() => <App />}
+  </Provider>,
   document.getElementById('root')
 );
