+++ examples/counter/index.js
@@ -1,7 +1,13 @@
 import React from 'react';
-import Root from './containers/Root';
+import { Provider } from 'react-redux';
+import App from './containers/App';
+import configureStore from './store/configureStore';
+
+const store = configureStore();
 
 React.render(
-  <Root />,
+  <Provider store={store}>
+    {() => <App />}
+  </Provider>,
   document.getElementById('root')
 );
