+++ examples/real-world/store/configureStore.js
@@ -1,17 +1,25 @@
-import { createStore, applyMiddleware, combineReducers } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import apiMiddleware from '../middleware/api';
 import loggerMiddleware from 'redux-logger';
-import * as reducers from '../reducers';
+import rootReducer from '../reducers';
 
-const reducer = combineReducers(reducers);
 const createStoreWithMiddleware = applyMiddleware(
   thunkMiddleware,
   apiMiddleware,
   loggerMiddleware
 )(createStore);
 
-// Creates a preconfigured store for this example.
 export default function configureStore(initialState) {
-  return createStoreWithMiddleware(reducer, initialState);
+  const store = createStoreWithMiddleware(rootReducer, initialState);
+
+  if (module.hot) {
+    // Enable Webpack hot module replacement for reducers
+    module.hot.accept('../reducers', () => {
+      const nextRootReducer = require('../reducers');
+      store.replaceReducer(nextRootReducer);
+    });
+  }
+
+  return store;
 }
