--- examples/todomvc/containers/Root.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import TodoApp from './TodoApp';
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import rootReducer from '../reducers';
-
-const store = createStore(rootReducer);
-
-export default class Root extends Component {
-  render() {
-    return (
-      <Provider store={store}>
-        {() => <TodoApp /> }
-      </Provider>
-    );
-  }
-}
