+++ examples/todomvc/index.js
@@ -1,10 +1,16 @@
 import 'babel-core/polyfill';
 
 import React from 'react';
-import Root from './containers/Root';
+import { Provider } from 'react-redux';
+import App from './containers/App';
+import configureStore from './store/configureStore';
 import 'todomvc-app-css/index.css';
 
+const store = configureStore();
+
 React.render(
-  <Root />,
+  <Provider store={store}>
+    {() => <App />}
+  </Provider>,
   document.getElementById('root')
 );
