+++ examples/todomvc/store/configureStore.js
@@ -1,13 +1,8 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import reducer from '../reducers';
-
-const createStoreWithMiddleware = applyMiddleware(
-  thunk
-)(createStore);
+import { createStore } from 'redux';
+import rootReducer from '../reducers';
 
 export default function configureStore(initialState) {
-  const store = createStoreWithMiddleware(reducer, initialState);
+  const store = createStore(rootReducer, initialState);
 
   if (module.hot) {
     // Enable Webpack hot module replacement for reducers
